Refactor format_bots module structure and imports
Moved format_bots out of make_bots, updating all related imports across the codebase. Adjusted test imports accordingly and removed an obsolete test file. This improves module organization and simplifies import paths.
