+++ ArWikiCats/format_bots/relation_mapping.py
