+++ ArWikiCats/ma_bots2/ar_lab/ar_lab_bot.py
@@ -11,7 +11,7 @@ from typing import Tuple
 from ...helps import logger
 from ...ma_bots2.year_or_typeo.bot_lab import label_for_startwith_year_or_typeo
 from ...ma_bots.country_bot import event2_d2
-from ...make_bots.format_bots.relation_mapping import translation_category_relations
+from ...format_bots.relation_mapping import translation_category_relations
 from ...make_bots.lazy_data_bots.bot_2018 import get_pop_All_18
 from ...make_bots.matables_bots.check_bot import check_key_new_players
 from ...make_bots.matables_bots.data import Keep_it_frist, Keep_it_last
