+++ ArWikiCats/ma_bots2/country2_bots/country2_label_bot.py
@@ -11,7 +11,7 @@ from ...old_bots.sports_bots import team_work
 
 from ...helps import logger
 from ...ma_bots import country_bot
-from ...make_bots.format_bots.relation_mapping import translation_category_relations
+from ...format_bots.relation_mapping import translation_category_relations
 from ...make_bots.lazy_data_bots.bot_2018 import get_pop_All_18
 from ...make_bots.matables_bots.bot import add_to_Films_O_TT
 from ...make_bots.matables_bots.check_bot import check_key_new_players
