+++ ArWikiCats/ma_bots2/year_or_typeo/reg_result.py
@@ -2,7 +2,7 @@ import functools
 import re
 from dataclasses import dataclass
 
-from ...make_bots.format_bots.relation_mapping import translation_category_relations
+from ...format_bots.relation_mapping import translation_category_relations
 
 
 @functools.lru_cache(maxsize=1)
