+++ ArWikiCats/make_bots/__init__.py
@@ -1,4 +1,4 @@
-from .format_bots import change_cat
+from ..format_bots import change_cat
 
 __all__ = [
     "change_cat",
