+++ ArWikiCats/make_bots/ends_keys.py
@@ -1,4 +1,4 @@
-from ...helps import len_print
+from ..helps import len_print
 
 # --- Tour de
