+++ ArWikiCats/old_bots/tmp_bot.py
@@ -12,7 +12,7 @@ from .sports_bots import team_work
 
 from ..helps import logger
 from ..ma_bots import ye_ts_bot
-from ..make_bots.format_bots.ends_keys import combined_suffix_mappings
+from ..make_bots.ends_keys import combined_suffix_mappings
 from ..make_bots.lazy_data_bots.bot_2018 import get_pop_All_18
 from ..make_bots.matables_bots.table1_bot import get_KAKO
 from ..make_bots.o_bots import parties_bot, univer
