+++ tests/make2_bots/format_bots/test_pf_keys_old.py
@@ -2,7 +2,7 @@
 """
 import pytest
 
-from ArWikiCats.make_bots.format_bots.pf_keys import change_key_mappings_replacements
+from ArWikiCats.format_bots.pf_keys import change_key_mappings_replacements
 
 data1 = {
     "africa cup of nations": "africa cup-of-nations",
