--- tests/make2_bots/o_bots/test_fax.py
@@ -1,29 +1,0 @@
-"""
-Tests
-"""
-
-import pytest
-
-from ArWikiCats.make_bots.languages_bot.languages_resolvers import te_language
-
-fast_data = {}
-
-
-@pytest.mark.parametrize("category, expected", fast_data.items(), ids=fast_data.keys())
-@pytest.mark.fast
-def test_fast_data(category: str, expected: str) -> None:
-    label = te_language(category)
-    assert label == expected
-
-
-def test_te_language() -> None:
-    # Test with a basic input
-    result = te_language("english language")
-    assert isinstance(result, str)
-
-    result_empty = te_language("")
-    assert isinstance(result_empty, str)
-
-    # Test with various inputs
-    result_various = te_language("french literature")
-    assert isinstance(result_various, str)
