+++ tests/utils/test_match_relation_word.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ArWikiCats.make_bots.format_bots.relation_mapping import translation_category_relations
+from ArWikiCats.format_bots.relation_mapping import translation_category_relations
 from ArWikiCats.utils.match_relation_word import get_relation_word, get_relation_word_new
 
 # ===============================================
